Refactor package structure for banking operations
Moved classes and DTOs related to banking operations under a more specific "operations" subpackage for better organization. Updated imports and references throughout the codebase to reflect the new structure. Enabled method-level security in SecurityConfig to enhance security configuration.
