+++ src/main/java/info/mackiewicz/bankapp/system/banking/operations/service/TransferOperationService.java
@@ -1,24 +1,23 @@
-package info.mackiewicz.bankapp.system.banking.service;
-
-import java.util.function.Supplier;
-
-import org.iban4j.Iban;
-import org.slf4j.MDC;
-import org.springframework.stereotype.Service;
+package info.mackiewicz.bankapp.system.banking.operations.service;
 
 import info.mackiewicz.bankapp.account.exception.AccountNotFoundByIbanException;
 import info.mackiewicz.bankapp.account.exception.AccountOwnershipException;
 import info.mackiewicz.bankapp.account.model.Account;
-import info.mackiewicz.bankapp.system.banking.api.dto.BankingOperationRequest;
-import info.mackiewicz.bankapp.system.banking.api.dto.TransferResponse;
-import info.mackiewicz.bankapp.system.banking.service.helpers.AccountSecurityService;
-import info.mackiewicz.bankapp.system.banking.service.helpers.TransactionBuildingService;
+import info.mackiewicz.bankapp.system.banking.operations.api.dto.BankingOperationRequest;
+import info.mackiewicz.bankapp.system.banking.operations.api.dto.TransferResponse;
+import info.mackiewicz.bankapp.system.banking.operations.service.helpers.AccountSecurityService;
+import info.mackiewicz.bankapp.system.banking.operations.service.helpers.TransactionBuildingService;
 import info.mackiewicz.bankapp.transaction.exception.TransactionBuildingException;
 import info.mackiewicz.bankapp.transaction.exception.TransactionValidationException;
 import info.mackiewicz.bankapp.transaction.model.Transaction;
 import info.mackiewicz.bankapp.transaction.service.TransactionService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.iban4j.Iban;
+import org.slf4j.MDC;
+import org.springframework.stereotype.Service;
+
+import java.util.function.Supplier;
 
 @Service
 @RequiredArgsConstructor
