+++ src/main/java/info/mackiewicz/bankapp/system/banking/operations/service/helpers/TransactionBuildingService.java
@@ -1,14 +1,14 @@
-package info.mackiewicz.bankapp.system.banking.service.helpers;
-
-import java.math.BigDecimal;
-
-import org.iban4j.Iban;
-import org.springframework.stereotype.Service;
+package info.mackiewicz.bankapp.system.banking.operations.service.helpers;
 
 import info.mackiewicz.bankapp.account.model.Account;
+import info.mackiewicz.bankapp.transaction.exception.TransactionBuildingException;
 import info.mackiewicz.bankapp.transaction.model.Transaction;
 import info.mackiewicz.bankapp.transaction.model.TransactionType;
 import lombok.RequiredArgsConstructor;
+import org.iban4j.Iban;
+import org.springframework.stereotype.Service;
+
+import java.math.BigDecimal;
 
 /**
  * Service for building transactions.
